Added funcionality delete at dashboard form

+++ src/app/views/dashboard/dashboard-form/dashboard-form.component.ts
@@ -1,14 +1,15 @@
 import { AfterViewInit, Component, OnInit, ViewChild } from '@angular/core';
 import { DateAdapter } from '@angular/material';
 import { BreakpointObserver } from '@angular/cdk/layout';
-import { TranslateService } from "@ngx-translate/core";
 import { MatPaginator } from "@angular/material/paginator";
 import { ActivatedRoute } from "@angular/router";
+import { MatTableDataSource } from '@angular/material/table';
+import { MatDialog } from "@angular/material/dialog";
+import { TranslateService } from "@ngx-translate/core";
 import { LocalStorageService } from "../../../core/services";
 import { DashboardFormService } from "./dashboard-form.service";
-import { MatDialog } from "@angular/material/dialog";
 import { DashboardReaderComponent } from "../dashboard-reader/dashboard-reader.component";
-import { MatTableDataSource } from '@angular/material/table';
+import Swal from 'sweetalert2';
 import * as _ from 'lodash';
 
 @Component({
@@ -68,11 +69,52 @@ export class DashboardFormComponent implements OnInit, AfterViewInit {
   removeProtocol(cod) {
     let protocols = JSON.parse(localStorage['protocols']);
 
-    _.remove(protocols, {'id': parseInt(cod)});
+    let protocol = _.filter(protocols, {'registration_code': parseInt(cod)});
 
-    localStorage.setItem('protocols', JSON.stringify(protocols));
+    Swal.fire({
+      title: 'Você tem certeza que quer deletar o '+ protocol[0].participant_name +'?',
+      text: "Você não consegue reverter essa ação depois!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim, pode deletar!'
+    }).then((result) => {
+      if (result.value) {
 
-    this.getProtocolReader();
+        if (protocol[0].id === null) {
+          _.remove(protocols, {'registration_code': parseInt(cod)});
+        }
+
+        if (protocol[0].id != null) {
+          _.remove(protocols, {'registration_code': parseInt(cod)});
+          protocol[0].active = 0;
+          let newProtocol = _.concat(protocols, protocol);
+          localStorage.setItem('protocols', JSON.stringify(newProtocol));
+          this.getProtocolReader();
+          Swal.fire(
+            'Deletado!',
+            'Candidato deletado com sucesso.',
+            'success'
+          );
+          this.synchronized = false;
+          this.localStorage.setItem('synchronized', JSON.stringify(this.synchronized));
+          return;
+        }
+
+        localStorage.setItem('protocols', JSON.stringify(protocols));
+        this.synchronized = false;
+        this.localStorage.setItem('synchronized', JSON.stringify(this.synchronized));
+
+        Swal.fire(
+          'Deletado!',
+          'Candidato deletado com sucesso.',
+          'success'
+        );
+
+        this.getProtocolReader();
+      }
+    });
   }
 
   private getProtocolReader(): void {
@@ -80,7 +122,7 @@ export class DashboardFormComponent implements OnInit, AfterViewInit {
 
     let group_id = this.route.snapshot.paramMap.get('group_id');
 
-    protocolReaderDataSource = _.filter(protocolReaderDataSource, {'group_reader_id': parseInt(group_id)});
+    protocolReaderDataSource = _.filter(protocolReaderDataSource, {'group_reader_id': parseInt(group_id), 'active': 1});
     this.dataSource = new MatTableDataSource(protocolReaderDataSource);
   }
